tempest-weather/requirements.txt
selenium
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
loguru==0.7.3
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
websockets==12.0
